requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.0.340
langchain-openai==0.0.2
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
python-multipart==0.0.6
googletrans==4.0.0rc1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
# Built-in modules (no installation needed)
# sqlite3, asyncio, json, datetime, typing, logging are built-in
